Dummy exception_traceback for activepapers.contents
This prevents a crash when ActivePapers such as the
unit tet framework are used from an external script
via activepapers.exploration.

Such a script stil crashes when exception_traceback
is called (this is hard to fix), but at least the
script doesn't already crash when importing a module
that imports exception_traceback.